Split out animals and plants into their own packages
