+++ pkg/animal/birds.go
@@ -1,0 +1,224 @@
+package animal
+
+func getBirds() []Animal {
+	birds := []Animal{
+		Animal{
+			Name:           "goose",
+			PluralName:     "geese",
+			AnimalType:     "bird",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "chicken",
+			PluralName:     "chickens",
+			AnimalType:     "bird",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "mudhen",
+			PluralName:     "mudhens",
+			AnimalType:     "bird",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "quail",
+			PluralName:     "quails",
+			AnimalType:     "bird",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "duck",
+			PluralName:     "ducks",
+			AnimalType:     "bird",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      true,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "flamingo",
+			PluralName:     "flamingos",
+			AnimalType:     "bird",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      true,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "eagle",
+			PluralName:     "eagles",
+			AnimalType:     "bird",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    true,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "falcon",
+			PluralName:     "falcons",
+			AnimalType:     "bird",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    true,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "hawk",
+			PluralName:     "hawks",
+			AnimalType:     "bird",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      false,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    true,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+	}
+
+	return birds
+}
