+++ pkg/animal/mammals.go
@@ -1,0 +1,824 @@
+package animal
+
+func getMammals() []Animal {
+	mammals := []Animal{
+		Animal{
+			Name:           "black bear",
+			PluralName:     "black bears",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "brown bear",
+			PluralName:     "brown bears",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "beaver",
+			PluralName:     "beavers",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      true,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "deer",
+			PluralName:     "deer",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      true,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "squirrel",
+			PluralName:     "squirrels",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "wolf",
+			PluralName:     "wolves",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "camel",
+			PluralName:     "camels",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        true,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    6,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "bison",
+			PluralName:     "bison",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "bobcat",
+			PluralName:     "bobcats",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "coyote",
+			PluralName:     "coyotes",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    true,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "horse",
+			PluralName:     "horses",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        true,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "cow",
+			PluralName:     "cows",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "elephant",
+			PluralName:     "elephants",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      true,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        true,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "jaguar",
+			PluralName:     "jaguars",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "tiger",
+			PluralName:     "tigers",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "goat",
+			PluralName:     "goats",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "sheep",
+			PluralName:     "sheep",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      true,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "cougar",
+			PluralName:     "cougars",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "alpaca",
+			PluralName:     "alpacas",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      true,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 4,
+			MaxTemperature: 7,
+		},
+		Animal{
+			Name:           "llama",
+			PluralName:     "llamas",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      true,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 7,
+		},
+		Animal{
+			Name:           "leopard",
+			PluralName:     "leopards",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 7,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "cheetah",
+			PluralName:     "cheetahs",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "lion",
+			PluralName:     "lions",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "hippo",
+			PluralName:     "hippos",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      true,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "zebra",
+			PluralName:     "zebras",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        true,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 7,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "antelope",
+			PluralName:     "antelopes",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      true,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "gazelle",
+			PluralName:     "gazelles",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      true,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "donkey",
+			PluralName:     "donkeys",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        true,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "rabbit",
+			PluralName:     "rabbits",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "ermine",
+			PluralName:     "ermines",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       true,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "mink",
+			PluralName:     "minks",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       true,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "fox",
+			PluralName:     "foxes",
+			AnimalType:     "mammal",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       true,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 6,
+		},
+		Animal{
+			Name:           "raccoon",
+			PluralName:     "raccoons",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      false,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    true,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 7,
+		},
+		Animal{
+			Name:           "reindeer",
+			PluralName:     "reindeers",
+			AnimalType:     "mammal",
+			EatsAnimals:    false,
+			EatsPlants:     true,
+			GivesBone:      true,
+			GivesEggs:      false,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      true,
+			GivesMeat:      true,
+			GivesMilk:      true,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   true,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 0,
+			MaxTemperature: 4,
+		},
+	}
+
+	return mammals
+}
