+++ pkg/animal/reptiles.go
@@ -1,0 +1,224 @@
+package animal
+
+func getReptiles() []Animal {
+	reptiles := []Animal{
+		Animal{
+			Name:           "alligator",
+			PluralName:     "alligators",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      true,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "crocodile",
+			PluralName:     "crocodiles",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      true,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "cobra",
+			PluralName:     "cobras",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     true,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "asp",
+			PluralName:     "asps",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     true,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "rattlesnake",
+			PluralName:     "rattlesnakes",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     true,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "black mamba",
+			PluralName:     "black mambas",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     true,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "coral snake",
+			PluralName:     "coral snakes",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     true,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "boa constrictor",
+			PluralName:     "boa constrictors",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     false,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+		Animal{
+			Name:           "monitor lizard",
+			PluralName:     "monitor lizards",
+			AnimalType:     "reptile",
+			EatsAnimals:    true,
+			EatsPlants:     false,
+			GivesBone:      false,
+			GivesEggs:      true,
+			GivesFur:       false,
+			GivesHide:      true,
+			GivesHorn:      false,
+			GivesMeat:      true,
+			GivesMilk:      false,
+			GivesWool:      false,
+			IsAquatic:      false,
+			IsMount:        false,
+			IsPackAnimal:   false,
+			IsScavenger:    false,
+			IsVenomous:     true,
+			MinHumidity:    0,
+			MaxHumidity:    10,
+			MinTemperature: 5,
+			MaxTemperature: 10,
+		},
+	}
+
+	return reptiles
+}
