+++ pkg/climate/animals.go
@@ -1,35 +1,11 @@
 package climate
 
-import "math/rand"
+import (
+	"github.com/ironarachne/world/pkg/animal"
+)
 
-// Animal is an animal
-type Animal struct {
-	Name           string
-	PluralName     string
-	AnimalType     string
-	EatsAnimals    bool
-	EatsPlants     bool
-	GivesBone      bool
-	GivesEggs      bool
-	GivesFur       bool
-	GivesHide      bool
-	GivesHorn      bool
-	GivesMeat      bool
-	GivesMilk      bool
-	GivesWool      bool
-	IsAquatic      bool
-	IsMount        bool
-	IsPackAnimal   bool
-	IsScavenger    bool
-	IsVenomous     bool
-	MinHumidity    int
-	MaxHumidity    int
-	MinTemperature int
-	MaxTemperature int
-}
-
-func (climate Climate) getFilteredAnimals() []Animal {
-	animals := getAllAnimals()
+func (climate Climate) getFilteredAnimals() []animal.Animal {
+	animals := animal.All()
 	animals = filterAnimalsForHumidity(climate.Humidity, animals)
 	animals = filterAnimalsForTemperature(climate.Temperature, animals)
 	animals = climate.filterAnimalsForWater(animals)
@@ -37,38 +13,8 @@ func (climate Climate) getFilteredAnimals() []Animal {
 	return animals
 }
 
-func getRandomAnimals(amount int, from []Animal) []Animal {
-	var animal Animal
-
-	animals := []Animal{}
-
-	if amount > len(from) {
-		amount = len(from)
-	}
-
-	for i := 0; i < amount; i++ {
-		animal = from[rand.Intn(len(from)-1)]
-		if !isAnimalInSlice(animal, animals) {
-			animals = append(animals, animal)
-		}
-	}
-
-	return animals
-}
-
-func isAnimalInSlice(animal Animal, animals []Animal) bool {
-	isIt := false
-	for _, a := range animals {
-		if a.Name == animal.Name {
-			isIt = true
-		}
-	}
-
-	return isIt
-}
-
-func filterAnimalsForHumidity(humidity int, animals []Animal) []Animal {
-	var filteredAnimals []Animal
+func filterAnimalsForHumidity(humidity int, animals []animal.Animal) []animal.Animal {
+	var filteredAnimals []animal.Animal
 
 	for _, a := range animals {
 		if a.MinHumidity <= humidity && a.MaxHumidity >= humidity {
@@ -79,8 +25,8 @@ func filterAnimalsForHumidity(humidity int, animals []Animal) []Animal {
 	return filteredAnimals
 }
 
-func filterAnimalsForTemperature(temperature int, animals []Animal) []Animal {
-	var filteredAnimals []Animal
+func filterAnimalsForTemperature(temperature int, animals []animal.Animal) []animal.Animal {
+	var filteredAnimals []animal.Animal
 
 	for _, a := range animals {
 		if a.MinTemperature <= temperature && a.MaxTemperature >= temperature {
@@ -91,8 +37,8 @@ func filterAnimalsForTemperature(temperature int, animals []Animal) []Animal {
 	return filteredAnimals
 }
 
-func (climate Climate) filterAnimalsForWater(animals []Animal) []Animal {
-	var filteredAnimals []Animal
+func (climate Climate) filterAnimalsForWater(animals []animal.Animal) []animal.Animal {
+	var filteredAnimals []animal.Animal
 
 	for _, a := range animals {
 		if a.IsAquatic && (climate.HasLakes || climate.HasOcean || climate.HasRivers || climate.HasWetlands) {
@@ -103,1286 +49,4 @@ func (climate Climate) filterAnimalsForWater(animals []Animal) []Animal {
 	}
 
 	return filteredAnimals
-}
-
-func getAllAnimals() []Animal {
-	var animals []Animal
-
-	birds := getBirds()
-	animals = append(animals, birds...)
-	mammals := getMammals()
-	animals = append(animals, mammals...)
-	reptiles := getReptiles()
-	animals = append(animals, reptiles...)
-
-	return animals
-}
-
-func getBirds() []Animal {
-	birds := []Animal{
-		Animal{
-			Name:           "goose",
-			PluralName:     "geese",
-			AnimalType:     "bird",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "chicken",
-			PluralName:     "chickens",
-			AnimalType:     "bird",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "mudhen",
-			PluralName:     "mudhens",
-			AnimalType:     "bird",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "quail",
-			PluralName:     "quails",
-			AnimalType:     "bird",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "duck",
-			PluralName:     "ducks",
-			AnimalType:     "bird",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      true,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "flamingo",
-			PluralName:     "flamingos",
-			AnimalType:     "bird",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      true,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "eagle",
-			PluralName:     "eagles",
-			AnimalType:     "bird",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    true,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "falcon",
-			PluralName:     "falcons",
-			AnimalType:     "bird",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    true,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "hawk",
-			PluralName:     "hawks",
-			AnimalType:     "bird",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      false,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    true,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-	}
-
-	return birds
-}
-
-func getMammals() []Animal {
-	mammals := []Animal{
-		Animal{
-			Name:           "black bear",
-			PluralName:     "black bears",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "brown bear",
-			PluralName:     "brown bears",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "beaver",
-			PluralName:     "beavers",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      true,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "deer",
-			PluralName:     "deer",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      true,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "squirrel",
-			PluralName:     "squirrels",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "wolf",
-			PluralName:     "wolves",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "camel",
-			PluralName:     "camels",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        true,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    6,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "bison",
-			PluralName:     "bison",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "bobcat",
-			PluralName:     "bobcats",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "coyote",
-			PluralName:     "coyotes",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    true,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "horse",
-			PluralName:     "horses",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        true,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "cow",
-			PluralName:     "cows",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "elephant",
-			PluralName:     "elephants",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      true,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        true,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "jaguar",
-			PluralName:     "jaguars",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "tiger",
-			PluralName:     "tigers",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "goat",
-			PluralName:     "goats",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "sheep",
-			PluralName:     "sheep",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      true,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "cougar",
-			PluralName:     "cougars",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "alpaca",
-			PluralName:     "alpacas",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      true,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 4,
-			MaxTemperature: 7,
-		},
-		Animal{
-			Name:           "llama",
-			PluralName:     "llamas",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      true,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 7,
-		},
-		Animal{
-			Name:           "leopard",
-			PluralName:     "leopards",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 7,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "cheetah",
-			PluralName:     "cheetahs",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "lion",
-			PluralName:     "lions",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "hippo",
-			PluralName:     "hippos",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      true,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "zebra",
-			PluralName:     "zebras",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        true,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 7,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "antelope",
-			PluralName:     "antelopes",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      true,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "gazelle",
-			PluralName:     "gazelles",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      true,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "donkey",
-			PluralName:     "donkeys",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        true,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "rabbit",
-			PluralName:     "rabbits",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "ermine",
-			PluralName:     "ermines",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       true,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "mink",
-			PluralName:     "minks",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       true,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "fox",
-			PluralName:     "foxes",
-			AnimalType:     "mammal",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       true,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 6,
-		},
-		Animal{
-			Name:           "raccoon",
-			PluralName:     "raccoons",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      false,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    true,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 7,
-		},
-		Animal{
-			Name:           "reindeer",
-			PluralName:     "reindeers",
-			AnimalType:     "mammal",
-			EatsAnimals:    false,
-			EatsPlants:     true,
-			GivesBone:      true,
-			GivesEggs:      false,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      true,
-			GivesMeat:      true,
-			GivesMilk:      true,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   true,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 4,
-		},
-	}
-
-	return mammals
-}
-
-func getReptiles() []Animal {
-	reptiles := []Animal{
-		Animal{
-			Name:           "alligator",
-			PluralName:     "alligators",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      true,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "crocodile",
-			PluralName:     "crocodiles",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      true,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "cobra",
-			PluralName:     "cobras",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     true,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "asp",
-			PluralName:     "asps",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     true,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "rattlesnake",
-			PluralName:     "rattlesnakes",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     true,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "black mamba",
-			PluralName:     "black mambas",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     true,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "coral snake",
-			PluralName:     "coral snakes",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     true,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "boa constrictor",
-			PluralName:     "boa constrictors",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-		Animal{
-			Name:           "monitor lizard",
-			PluralName:     "monitor lizards",
-			AnimalType:     "reptile",
-			EatsAnimals:    true,
-			EatsPlants:     false,
-			GivesBone:      false,
-			GivesEggs:      true,
-			GivesFur:       false,
-			GivesHide:      true,
-			GivesHorn:      false,
-			GivesMeat:      true,
-			GivesMilk:      false,
-			GivesWool:      false,
-			IsAquatic:      false,
-			IsMount:        false,
-			IsPackAnimal:   false,
-			IsScavenger:    false,
-			IsVenomous:     true,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 5,
-			MaxTemperature: 10,
-		},
-	}
-
-	return reptiles
 }
