+++ pkg/climate/plants.go
@@ -1,31 +1,14 @@
 package climate
 
-import "math/rand"
+import (
+	"github.com/ironarachne/world/pkg/plant"
+)
 
-// Plant is a plant
-type Plant struct {
-	Name           string
-	PluralName     string
-	IsFiber        bool
-	IsFruit        bool
-	IsGrain        bool
-	IsHerb         bool
-	IsMedicine     bool
-	IsNut          bool
-	IsRoot         bool
-	IsSpice        bool
-	IsToxic        bool
-	IsVegetable    bool
-	MinHumidity    int
-	MaxHumidity    int
-	MinTemperature int
-	MaxTemperature int
-}
-
-func (climate Climate) getFilteredPlants() []Plant {
-	plants := getAllPlants()
+func (climate Climate) getFilteredPlants() []plant.Plant {
+	plants := plant.All()
 	plants = filterPlantsForHumidity(climate.Humidity, plants)
 	plants = filterPlantsForTemperature(climate.Temperature, plants)
+	plants = append(plants, plant.RandomFiber())
 
 	if len(plants) < 1 {
 		panic("Found no plants for climate " + climate.Name)
@@ -34,52 +17,8 @@ func (climate Climate) getFilteredPlants() []Plant {
 	return plants
 }
 
-func getRandomFiber(from []Plant) Plant {
-	var fibers []Plant
-
-	for _, p := range from {
-		if p.IsFiber {
-			fibers = append(fibers, p)
-		}
-	}
-
-	return fibers[rand.Intn(len(fibers))]
-}
-
-func getRandomPlants(amount int, from []Plant) []Plant {
-	var plant Plant
-
-	plants := []Plant{}
-
-	if amount > len(from) {
-		amount = len(from)
-	}
-
-	plants = append(plants, getRandomFiber(from))
-
-	for i := 1; i < amount; i++ {
-		plant = from[rand.Intn(len(from))]
-		if !isPlantInSlice(plant, plants) {
-			plants = append(plants, plant)
-		}
-	}
-
-	return plants
-}
-
-func isPlantInSlice(plant Plant, plants []Plant) bool {
-	isIt := false
-	for _, a := range plants {
-		if a.Name == plant.Name {
-			isIt = true
-		}
-	}
-
-	return isIt
-}
-
-func filterPlantsForHumidity(humidity int, plants []Plant) []Plant {
-	var filteredPlants []Plant
+func filterPlantsForHumidity(humidity int, plants []plant.Plant) []plant.Plant {
+	var filteredPlants []plant.Plant
 
 	for _, a := range plants {
 		if a.MinHumidity <= humidity && a.MaxHumidity >= humidity {
@@ -90,8 +29,8 @@ func filterPlantsForHumidity(humidity int, plants []Plant) []Plant {
 	return filteredPlants
 }
 
-func filterPlantsForTemperature(temperature int, plants []Plant) []Plant {
-	var filteredPlants []Plant
+func filterPlantsForTemperature(temperature int, plants []plant.Plant) []plant.Plant {
+	var filteredPlants []plant.Plant
 
 	for _, a := range plants {
 		if a.MinTemperature <= temperature && a.MaxTemperature >= temperature {
@@ -100,1298 +39,4 @@ func filterPlantsForTemperature(temperature int, plants []Plant) []Plant {
 	}
 
 	return filteredPlants
-}
-
-func getAllPlants() []Plant {
-	var plants []Plant
-
-	fibers := getFibers()
-	plants = append(plants, fibers...)
-	fruits := getFruits()
-	plants = append(plants, fruits...)
-	grains := getGrains()
-	plants = append(plants, grains...)
-	herbs := getHerbs()
-	plants = append(plants, herbs...)
-	vegetables := getVegetables()
-	plants = append(plants, vegetables...)
-
-	return plants
-}
-
-func getFibers() []Plant {
-	fibers := []Plant{
-		Plant{
-			Name:           "cotton",
-			PluralName:     "cotton",
-			IsFiber:        true,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "flax",
-			PluralName:     "flax",
-			IsFiber:        true,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 0,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "hemp",
-			PluralName:     "hemp",
-			IsFiber:        true,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 9,
-		},
-		Plant{
-			Name:           "coir",
-			PluralName:     "coir",
-			IsFiber:        true,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    6,
-			MaxHumidity:    10,
-			MinTemperature: 8,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "papyrus",
-			PluralName:     "papyrus",
-			IsFiber:        true,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    0,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 9,
-		},
-		Plant{
-			Name:           "jute",
-			PluralName:     "jute",
-			IsFiber:        true,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 9,
-		},
-	}
-
-	return fibers
-}
-
-func getFruits() []Plant {
-	fruits := []Plant{
-		Plant{
-			Name:           "blackberry",
-			PluralName:     "blackberries",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "blueberry",
-			PluralName:     "blueberries",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "raspberry",
-			PluralName:     "raspberries",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "strawberry",
-			PluralName:     "strawberries",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "boysenberry",
-			PluralName:     "boysenberries",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "gooseberry",
-			PluralName:     "gooseberries",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "avocado",
-			PluralName:     "avocados",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    6,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "cantaloupe",
-			PluralName:     "cantaloupes",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "coconut",
-			PluralName:     "coconuts",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "dragonfruit",
-			PluralName:     "dragonfruits",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "honeydew",
-			PluralName:     "honeydews",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "watermelon",
-			PluralName:     "watermelons",
-			IsFiber:        false,
-			IsFruit:        true,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-	}
-
-	return fruits
-}
-
-func getGrains() []Plant {
-	grains := []Plant{
-		Plant{
-			Name:           "rye",
-			PluralName:     "ryes",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        true,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    7,
-			MinTemperature: 3,
-			MaxTemperature: 7,
-		},
-		Plant{
-			Name:           "oat",
-			PluralName:     "oats",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        true,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    7,
-			MinTemperature: 3,
-			MaxTemperature: 7,
-		},
-		Plant{
-			Name:           "millet",
-			PluralName:     "millets",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        true,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    7,
-			MinTemperature: 3,
-			MaxTemperature: 7,
-		},
-		Plant{
-			Name:           "barley",
-			PluralName:     "barleys",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        true,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    7,
-			MinTemperature: 3,
-			MaxTemperature: 7,
-		},
-		Plant{
-			Name:           "wheat",
-			PluralName:     "wheats",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        true,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    7,
-			MinTemperature: 3,
-			MaxTemperature: 7,
-		},
-		Plant{
-			Name:           "rice",
-			PluralName:     "rices",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        true,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    6,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-	}
-
-	return grains
-}
-
-func getHerbs() []Plant {
-	herbs := []Plant{
-		Plant{
-			Name:           "parsley",
-			PluralName:     "parsleys",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "brahmi",
-			PluralName:     "brahmis",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     true,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "basil",
-			PluralName:     "basils",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "cilantro",
-			PluralName:     "cilantros",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "saffron",
-			PluralName:     "saffrons",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 6,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "rosemary",
-			PluralName:     "rosemarys",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "thyme",
-			PluralName:     "thymes",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "black pepper",
-			PluralName:     "black peppers",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "oregano",
-			PluralName:     "oreganos",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "mint",
-			PluralName:     "mints",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "lemongrass",
-			PluralName:     "lemongrasses",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 7,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "coriander",
-			PluralName:     "corianders",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "clove",
-			PluralName:     "cloves",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "sage",
-			PluralName:     "sages",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 3,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "turmeric",
-			PluralName:     "turmerics",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 7,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "vanilla",
-			PluralName:     "vanillas",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 7,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "galangal",
-			PluralName:     "galangals",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 7,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "ginger",
-			PluralName:     "gingers",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         true,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    false,
-			MinHumidity:    5,
-			MaxHumidity:    10,
-			MinTemperature: 7,
-			MaxTemperature: 10,
-		},
-	}
-
-	return herbs
-}
-
-func getVegetables() []Plant {
-	vegetables := []Plant{
-		Plant{
-			Name:           "potato",
-			PluralName:     "potatoes",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         true,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "onion",
-			PluralName:     "onions",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         true,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "beet",
-			PluralName:     "beets",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         true,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "radish",
-			PluralName:     "radishes",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         true,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "yam",
-			PluralName:     "yams",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         true,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "carrot",
-			PluralName:     "carrots",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         true,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "eggplant",
-			PluralName:     "eggplants",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "kohlrabi",
-			PluralName:     "kohlrabis",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "cabbage",
-			PluralName:     "cabbages",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "lettuce",
-			PluralName:     "lettuces",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "fiddlehead",
-			PluralName:     "fiddleheads",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "broccoli",
-			PluralName:     "broccolis",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "cauliflower",
-			PluralName:     "cauliflowers",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "tomato",
-			PluralName:     "tomatoes",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "bell pepper",
-			PluralName:     "bell peppers",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "bok choy",
-			PluralName:     "bok choys",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "leek",
-			PluralName:     "leeks",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "brussel sprout",
-			PluralName:     "brussel sprouts",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "green bean",
-			PluralName:     "green beans",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "black bean",
-			PluralName:     "black beans",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "pinto bean",
-			PluralName:     "pinto beans",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "lima bean",
-			PluralName:     "lima beans",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "kidney bean",
-			PluralName:     "kidney beans",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "corn",
-			PluralName:     "corns",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "jalapeno",
-			PluralName:     "jalapenos",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "habanero",
-			PluralName:     "habaneros",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-		Plant{
-			Name:           "serrano",
-			PluralName:     "serranos",
-			IsFiber:        false,
-			IsFruit:        false,
-			IsGrain:        false,
-			IsHerb:         false,
-			IsMedicine:     false,
-			IsNut:          false,
-			IsRoot:         false,
-			IsSpice:        false,
-			IsToxic:        false,
-			IsVegetable:    true,
-			MinHumidity:    3,
-			MaxHumidity:    10,
-			MinTemperature: 2,
-			MaxTemperature: 10,
-		},
-	}
-
-	return vegetables
 }
