+++ pkg/plant/fruit.go
@@ -1,0 +1,224 @@
+package plant
+
+func getFruits() []Plant {
+	fruits := []Plant{
+		Plant{
+			Name:           "blackberry",
+			PluralName:     "blackberries",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "blueberry",
+			PluralName:     "blueberries",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "raspberry",
+			PluralName:     "raspberries",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "strawberry",
+			PluralName:     "strawberries",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "boysenberry",
+			PluralName:     "boysenberries",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "gooseberry",
+			PluralName:     "gooseberries",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "avocado",
+			PluralName:     "avocados",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    6,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "cantaloupe",
+			PluralName:     "cantaloupes",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "coconut",
+			PluralName:     "coconuts",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "dragonfruit",
+			PluralName:     "dragonfruits",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "honeydew",
+			PluralName:     "honeydews",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "watermelon",
+			PluralName:     "watermelons",
+			IsFiber:        false,
+			IsFruit:        true,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+	}
+
+	return fruits
+}
