+++ pkg/plant/grains.go
@@ -1,0 +1,116 @@
+package plant
+
+func getGrains() []Plant {
+	grains := []Plant{
+		Plant{
+			Name:           "rye",
+			PluralName:     "ryes",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        true,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    7,
+			MinTemperature: 3,
+			MaxTemperature: 7,
+		},
+		Plant{
+			Name:           "oat",
+			PluralName:     "oats",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        true,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    7,
+			MinTemperature: 3,
+			MaxTemperature: 7,
+		},
+		Plant{
+			Name:           "millet",
+			PluralName:     "millets",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        true,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    7,
+			MinTemperature: 3,
+			MaxTemperature: 7,
+		},
+		Plant{
+			Name:           "barley",
+			PluralName:     "barleys",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        true,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    7,
+			MinTemperature: 3,
+			MaxTemperature: 7,
+		},
+		Plant{
+			Name:           "wheat",
+			PluralName:     "wheats",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        true,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    7,
+			MinTemperature: 3,
+			MaxTemperature: 7,
+		},
+		Plant{
+			Name:           "rice",
+			PluralName:     "rices",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        true,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    6,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+	}
+
+	return grains
+}
