+++ pkg/plant/herbs.go
@@ -1,0 +1,332 @@
+package plant
+
+func getHerbs() []Plant {
+	herbs := []Plant{
+		Plant{
+			Name:           "parsley",
+			PluralName:     "parsleys",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "brahmi",
+			PluralName:     "brahmis",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     true,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "basil",
+			PluralName:     "basils",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "cilantro",
+			PluralName:     "cilantros",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "saffron",
+			PluralName:     "saffrons",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 6,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "rosemary",
+			PluralName:     "rosemarys",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "thyme",
+			PluralName:     "thymes",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "black pepper",
+			PluralName:     "black peppers",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "oregano",
+			PluralName:     "oreganos",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "mint",
+			PluralName:     "mints",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "lemongrass",
+			PluralName:     "lemongrasses",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 7,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "coriander",
+			PluralName:     "corianders",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "clove",
+			PluralName:     "cloves",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "sage",
+			PluralName:     "sages",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 3,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "turmeric",
+			PluralName:     "turmerics",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 7,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "vanilla",
+			PluralName:     "vanillas",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 7,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "galangal",
+			PluralName:     "galangals",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 7,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "ginger",
+			PluralName:     "gingers",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         true,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    false,
+			MinHumidity:    5,
+			MaxHumidity:    10,
+			MinTemperature: 7,
+			MaxTemperature: 10,
+		},
+	}
+
+	return herbs
+}
