+++ pkg/plant/vegetables.go
@@ -1,0 +1,494 @@
+package plant
+
+func getVegetables() []Plant {
+	vegetables := []Plant{
+		Plant{
+			Name:           "potato",
+			PluralName:     "potatoes",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         true,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "onion",
+			PluralName:     "onions",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         true,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "beet",
+			PluralName:     "beets",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         true,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "radish",
+			PluralName:     "radishes",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         true,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "yam",
+			PluralName:     "yams",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         true,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "carrot",
+			PluralName:     "carrots",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         true,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "eggplant",
+			PluralName:     "eggplants",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "kohlrabi",
+			PluralName:     "kohlrabis",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "cabbage",
+			PluralName:     "cabbages",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "lettuce",
+			PluralName:     "lettuces",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "fiddlehead",
+			PluralName:     "fiddleheads",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "broccoli",
+			PluralName:     "broccolis",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "cauliflower",
+			PluralName:     "cauliflowers",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "tomato",
+			PluralName:     "tomatoes",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "bell pepper",
+			PluralName:     "bell peppers",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "bok choy",
+			PluralName:     "bok choys",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "leek",
+			PluralName:     "leeks",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "brussel sprout",
+			PluralName:     "brussel sprouts",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "green bean",
+			PluralName:     "green beans",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "black bean",
+			PluralName:     "black beans",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "pinto bean",
+			PluralName:     "pinto beans",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "lima bean",
+			PluralName:     "lima beans",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "kidney bean",
+			PluralName:     "kidney beans",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "corn",
+			PluralName:     "corns",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "jalapeno",
+			PluralName:     "jalapenos",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "habanero",
+			PluralName:     "habaneros",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+		Plant{
+			Name:           "serrano",
+			PluralName:     "serranos",
+			IsFiber:        false,
+			IsFruit:        false,
+			IsGrain:        false,
+			IsHerb:         false,
+			IsMedicine:     false,
+			IsNut:          false,
+			IsRoot:         false,
+			IsSpice:        false,
+			IsToxic:        false,
+			IsVegetable:    true,
+			MinHumidity:    3,
+			MaxHumidity:    10,
+			MinTemperature: 2,
+			MaxTemperature: 10,
+		},
+	}
+
+	return vegetables
+}
